Add support for Mongoid 'legacy' (Rails 2.3.*) branch
This patch allows Devise's v1.0 branch to be used with the current
'legacy' branch of Mongoid and Rails 2.3.5.  This is useful for Rails
2.3 projects using MongoMapper and Devise, because it allows to first
switch to Mongoid, and _then_ to Rails 3 in two separate stages.

All unit tests pass, with one small tweak to remove milliseconds from
Mongoid's created_at timestamps.

Note that we need to do some ugly patching of Mongoid's validation layer
(at test-time only) to make it give the same validation messages as
ActiveRecord.

The following Gemfile is enough to make both the ActiveRecord and
Mongoid unit tests pass, though you may need to downgrade your version
of 'mongo' and 'bson_ext' if you also want to test MongoMapper.

  gem 'mongo', '1.0.1'
  gem 'bson_ext', '1.0.1'
  gem 'mongoid', :git => 'git://github.com/durran/mongoid.git', :branch => 'legacy'
  gem 'rails', '2.3.5'
  gem 'warden', '~> 0.10.3'
  gem 'will_paginate', '~> 2.3.11'
  gem 'webrat'
  gem 'mocha'
  gem 'bcrypt-ruby'
  gem 'rack-test', :require => nil
  gem 'jeweler'
  gem 'sqlite3-ruby', :require => nil

Tests were run as:

  bundle exec rake test DEVISE_ORM=active_record
  bundle exec rake test DEVISE_ORM=mongoid

Note that some of this code is adapted from the existing MongoMapper
driver in Devise.